requirements.txt
# Whereas install_requires requirements are minimal, requirements files often contain an
# exhaustive listing of pinned versions for the purpose of achieving repeatable
# installations of a complete environment.

paddlepaddle==2.1.0
paddlenlp==2.0.2
PyYAML==5.3.1
pydantic==1.8.1